Mensajes Sweetalert2
Se agregaron mensajes animados,
asi como tambien se resolvio un bug en la busqueda (searchE)

+++ public/js/director/espacios.js
@@ -40,12 +40,23 @@ new Vue({
             this.$http.post(UrlEspacio,e)
             .then(function(json){
                 $('#agregar_espacio').modal('hide');
+                Swal.fire({
+                    position: 'center',
+                    icon: 'success',
+                    title: 'Guardado exitosamente',
+                    showConfirmButton: false,
+                    timer: 1500
+                })
                 this.nombre="";
                 this.ubicacion="";
                 this.cupo="";
                 this.getEspacio();
             }).catch(function(json){
-                console.log(json);
+                Swal.fire({
+                    icon: 'error',
+                    title: 'Ha ocurrido un error',
+                    text: 'No deje campos vacios',
+                })
             })
         },
         editarEspacio:function(id){
@@ -56,7 +67,6 @@ new Vue({
                 this.ubicacion = json.data.ubicacion
                 this.cupo = json.data.cupo
                 this.id_e = json.data.id_espacio
-
                 $('#agregar_espacio').modal('show');
             })
         },
@@ -69,24 +79,52 @@ new Vue({
             this.$http.patch(UrlEspacio + '/' + this.id_e,espac)
             .then(function(){
                 $('#agregar_espacio').modal('hide');
-                alert ('agregado con exito');
+                Swal.fire({
+                    position: 'center',
+                    icon: 'success',
+                    title: 'Guardado exitosamente',
+                    showConfirmButton: false,
+                    timer: 1500
+                })
                 this.getEspacio();
-
                 this.nombre="";
                 this.ubicacion="";
                 this.cupo="";
                 this.editar=false;
+            }).catch(function(){
+                Swal.fire({
+                    icon: 'error',
+                    title: 'Ha ocurrido un error',
+                    text: 'No deje campos vacios',
+                })
             })
         },
         eliminarEspacio:function(id){
-            this.$http.delete(UrlEspacio +'/'+id)
-            .then(function(){
-                alert ('eliminado con exito');
-                // this.getEspacio();
-            }).catch(function(json){
-                console.log(json);
+            Swal.fire({
+				title: "No podrás revertir este cambio!,¿Estás seguro?",
+				icon: 'warning',
+				showCancelButton: true,
+				confirmButtonColor: '#3085d6',
+				cancelButtonColor: '#d33',
+				confirmButtonText: 'Sí, bórralo',
+				cancelButtonText:'No, cancelar',
+			}).then((result)=>{
+                if(result.value){
+                    this.$http.delete(UrlEspacio +'/'+id)
+                    .then(function(){
+                        Swal.fire({
+                            position: 'center',
+                            icon: 'success',
+                            title: 'Eliminado exitosamente',
+                            showConfirmButton: false,
+                            timer: 1500
+                        })
+                        this.getEspacio();
+                    }).catch(function(json){
+                        console.log(json);
+                    })
+                }
             })
-
         },
         salir:function(){
             this.editar=false;
@@ -98,7 +136,9 @@ new Vue({
     computed:{
         searchE:function(){
             return this.espacio.filter((espac)=>{
-                return espac.nombre.match(this.search.trim().toLowerCase());
+                return espac.nombre.toLowerCase().match(this.search.trim().toLowerCase())
+                || espac.ubicacion.toLowerCase().match(this.search.trim().toLowerCase())
+                || espac.cupo.toLowerCase().match(this.search.trim().toLowerCase());
             })
         }
     }
